Added registration form to web app

+++ webapp/pages/login.js
@@ -23,12 +23,12 @@ function LoginPage() {
         event.preventDefault();
         var resp = await loginUser(email, pswd);
         
-        if (resp.status === 200 && resp.data.exists === true) {
+        if (resp && resp.status === 200 && resp.data.exists === true) {
 
             localStorage.setItem('token', resp.data.accessToken);
             // alert(`Logging in ${resp.data.user.email}!`);
             Router.push('/');
-        };
+        }
     };
 
     return (
@@ -41,14 +41,17 @@ function LoginPage() {
                 
                 <h4>Welcome back! Please sign in.</h4>
 
-                <label for="email">Email</label>
+                <label htmlFor="email">Email</label>
                 <input type='email' onChange={handleEChange} value={email} placeholder='Enter your email address' />
-                <label for="email">Password</label>
+                <label htmlFor="email">Password</label>
                 <input type='password' onChange={handlePChange} value={pswd} placeholder='Enter your password' />
                 
                 <button className={styles.athBtn}>Sign In</button>
             
             </form>
+            <Link href='/register'>
+                <a>Not a user yet? Join here.</a>
+            </Link>
         </>
     )
 }
